+++ Tests/Api/MultipleTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DBlackborough\Quill\Tests\Html;
+namespace DBlackborough\Quill\Tests\Api;
 
 use DBlackborough\Quill\RenderMultiple;
