+++ Tests/Combined/Html/BreaksTest.php
@@ -1,36 +1,39 @@
 <?php
 
-namespace DBlackborough\Quill\Tests\Composite\Markdown;
+namespace DBlackborough\Quill\Tests\Composite\Html;
 
 require __DIR__ . '../../../../vendor/autoload.php';
 
-use DBlackborough\Quill\Options;
 use DBlackborough\Quill\Render as QuillRender;
 
 /**
  * Paragraph tests
  */
-final class ParagraphTest extends \PHPUnit\Framework\TestCase
+final class BreaksTest extends \PHPUnit\Framework\TestCase
 {
     private $delta_paragraphs_with_attributes = '{"ops":[{"insert":"This is a three "},{"attributes":{"bold":true},"insert":"paragraph"},{"insert":" test\n\nthe "},{"attributes":{"strike":true},"insert":"difference"},{"insert":" being this time we \n\nare "},{"attributes":{"underline":true},"insert":"going to add"},{"insert":" attributes.\n"}]}';
     private $delta_single_paragraph = '{"ops":[{"insert":"Lorem ipsum dolor sit amet"}]}';
     private $delta_two_paragraphs = '{"ops":[{"insert":"Lorem ipsum dolor sit amet.\n\nLorem ipsum dolor sit amet."}]}';
     private $delta_three_paragraphs = '{"ops":[{"insert":"This is a single entry that \n\nshould create three paragraphs \n\nof HTML.\n"}]}';
+    private $delta_line_breaks = '{"ops":[{"insert":"Line 1, should have a BR\nLine 2, should have a BR\nLine 3\n"}]}';
+    private $delta_paragraph_then_line_breaks = '{"ops":[{"insert":"Paragraph\n\nLine 1, should have a BR\nLine 2, should have a BR\nLine 3\n"}]}';
 
-    private $expected_paragraphs_with_attributes = "This is a three **paragraph** test
-
-the difference being this time we 
-
-are going to add attributes.";
-    private $expected_single_paragraph = 'Lorem ipsum dolor sit amet';
-    private $expected_two_paragraphs = 'Lorem ipsum dolor sit amet.
-
-Lorem ipsum dolor sit amet.';
-    private $expected_three_paragraphs = 'This is a single entry that 
-
-should create three paragraphs 
-
-of HTML.';
+    private $expected_paragraphs_with_attributes = "<p>This is a three <strong>paragraph</strong> test</p>
+<p>the <s>difference</s> being this time we </p>
+<p>are <u>going to add</u> attributes.</p>";
+    private $expected_single_paragraph = '<p>Lorem ipsum dolor sit amet</p>';
+    private $expected_two_paragraphs = '<p>Lorem ipsum dolor sit amet.</p>
+<p>Lorem ipsum dolor sit amet.</p>';
+    private $expected_three_paragraphs = '<p>This is a single entry that </p>
+<p>should create three paragraphs </p>
+<p>of HTML.</p>';
+    private $expected_line_breaks = "<p>Line 1, should have a BR<br />
+Line 2, should have a BR<br />
+Line 3</p>";
+    private $expected_paragraph_then_line_breaks = "<p>Paragraph</p>
+<p>Line 1, should have a BR<br />
+Line 2, should have a BR<br />
+Line 3</p>";
 
     /**
      * Test paragraphs with attributes
@@ -43,10 +46,7 @@ of HTML.';
         $result = null;
 
         try {
-            $quill = new QuillRender(
-                $this->delta_paragraphs_with_attributes,
-                OPTIONS::FORMAT_MARKDOWN
-            );
+            $quill = new QuillRender($this->delta_paragraphs_with_attributes);
             $result = $quill->render();
         } catch (\Exception $e) {
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
@@ -70,10 +70,7 @@ of HTML.';
         $result = null;
 
         try {
-            $quill = new QuillRender(
-                $this->delta_single_paragraph,
-                OPTIONS::FORMAT_MARKDOWN
-            );
+            $quill = new QuillRender($this->delta_single_paragraph);
             $result = $quill->render();
         } catch (\Exception $e) {
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
@@ -97,10 +94,7 @@ of HTML.';
         $result = null;
 
         try {
-            $quill = new QuillRender(
-                $this->delta_two_paragraphs,
-                OPTIONS::FORMAT_MARKDOWN
-            );
+            $quill = new QuillRender($this->delta_two_paragraphs);
             $result = $quill->render();
         } catch (\Exception $e) {
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
@@ -124,10 +118,7 @@ of HTML.';
         $result = null;
 
         try {
-            $quill = new QuillRender(
-                $this->delta_three_paragraphs,
-                OPTIONS::FORMAT_MARKDOWN
-            );
+            $quill = new QuillRender($this->delta_three_paragraphs);
             $result = $quill->render();
         } catch (\Exception $e) {
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
@@ -137,6 +128,54 @@ of HTML.';
             $this->expected_three_paragraphs,
             trim($result),
             __METHOD__ . ' - Three paragraphs failure'
+        );
+    }
+
+    /**
+     * test two lines breaks inside a paragraph, no other attributes
+     *
+     * @return void
+     * @throws \Exception
+     */
+    public function testTwoLineBreaksInAParagraph()
+    {
+        $result = null;
+
+        try {
+            $quill = new QuillRender($this->delta_line_breaks);
+            $result = $quill->render();
+        } catch (\Exception $e) {
+            $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
+        }
+
+        $this->assertEquals(
+            $this->expected_line_breaks,
+            trim($result),
+            __METHOD__ . ' - Lines breaks in a simple paragraph, failure'
+        );
+    }
+
+    /**
+     * Test two lines breaks inside a paragraph, following a paragraph, no other attributes
+     *
+     * @return void
+     * @throws \Exception
+     */
+    public function testParagraphThenTwoLineBreaksInAParagraphFollowingParagraph()
+    {
+        $result = null;
+
+        try {
+            $quill = new QuillRender($this->delta_paragraph_then_line_breaks);
+            $result = $quill->render();
+        } catch (\Exception $e) {
+            $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
+        }
+
+        $this->assertEquals(
+            $this->expected_paragraph_then_line_breaks,
+            trim($result),
+            __METHOD__ . ' - Lines breaks in a simple paragraph, following a paragraph, failure'
         );
     }
 }
