+++ src/Delta/Html/Header.php
@@ -26,7 +26,7 @@ class Header extends Delta
         $this->insert = $insert;
         $this->attributes = $attributes;
 
-        $this->tag = Options::TAG_HEADER . $this->attributes['header'];
+        $this->tag = Options::HTML_TAG_HEADER . $this->attributes['header'];
     }
 
     /**
