+++ src/Delta/Markdown/Bold.php
@@ -23,7 +23,7 @@ class Bold extends Delta
      */
     public function __construct(string $insert, array $attributes = [])
     {
-        $this->token = Options::TOKEN_BOLD;
+        $this->token = Options::MARKDOWN_TOKEN_BOLD;
 
         $this->insert = $insert;
         $this->attributes = $attributes;
